+++ test/lib/zeitwerk/test_ignore.rb
@@ -2,6 +2,18 @@ require "test_helper"
 require "set"
 
 class TestIgnore < LoaderTest
+  def this_dir
+    @this_dir ||= __dir__
+  end
+
+  def this_file
+    @this_file ||= File.expand_path(__FILE__, this_dir)
+  end
+
+  def ascendant
+    @dir_up ||= File.expand_path("#{this_dir}/../..")
+  end
+
   test "ignored root directories are ignored" do
     files = [["x.rb", "X = true"]]
     with_files(files) do
@@ -135,5 +147,40 @@ class TestIgnore < LoaderTest
       assert Post
       assert_raises(NameError) { PostTest }
     end
+  end
+
+  test "returns true if a directory is ignored as is" do
+    loader.ignore(this_dir)
+    assert loader.ignores?(this_dir)
+  end
+
+  test "returns true if a file is ignored as is" do
+    loader.ignore(this_file)
+    assert loader.ignores?(this_file)
+  end
+
+  test "returns true for a descendant of an ignored directory" do
+    loader.ignore(ascendant)
+    assert loader.ignores?(this_dir)
+  end
+
+  test "returns true for a file in a descendant of an ignored directory" do
+    loader.ignore(ascendant)
+    assert loader.ignores?(this_file)
+  end
+
+  test "returns false for the directory of an ignored file" do
+    loader.ignore(this_file)
+    assert !loader.ignores?(this_dir)
+  end
+
+  test "returns false for an ascendant directory of an ignored directory" do
+    loader.ignore(this_dir)
+    assert !loader.ignores?(ascendant)
+  end
+
+  test "returns false if nothing is ignored" do
+    assert !loader.ignores?(this_dir)
+    assert !loader.ignores?(this_file)
   end
 end